Standardise pin controls; per-chip LED drivers; perf scenarios + 3-board sweep
Moves every single-GPIO control to the standard Pin control type, compiles each LED driver only into the chips whose silicon can run it, and replaces three overlapping perf scenarios with two incremental "add one subsystem, measure the delta" sweeps — then runs them across the classic ESP32, S3, and P4 to populate a three-board performance analysis. Also processes a second CodeRabbit batch.

KPI: 16384lights | PC:363KB | tick:118/99/118/9/1/317/35/16/18/114/10us(FPS:8474/10101/8474/111111/1000000/3154/28571/62500/55555/8771/100000) | src:97(19318) | test:66(9794) | lizard:74w
(ESP32 KPI tick omitted: the S3 was unreachable for a live render capture at commit time. This change is pin-control type, per-chip driver gating, scenarios and docs — no render-path code — and all three boards were verified rendering via the perf scenarios earlier this session.)

Core:
- AudioModule: wsPin / sdPin / sckPin moved from uint16/addUint16 to int8/addPin (the standard single-GPIO Pin control). Unset sentinel is now -1 (was 0), so GPIO 0 is a usable mic pin — a latent bug fix; the idle-when-unset guard switched from ==0 to <0.

Light domain:
- main.cpp: each LED driver's #include + registerType is now gated on the SOC capability macro it needs — RmtLedDriver on CONFIG_SOC_RMT_SUPPORTED, LcdLedDriver on CONFIG_SOC_LCDCAM_I80_LCD_SUPPORTED, ParlioLedDriver on CONFIG_SOC_PARLIO_SUPPORTED. A board's binary now carries only the drivers its chip can run (verified: the classic ELF has 0 LCD/Parlio symbols, 26 RMT; the classic type picker offers only RmtLedDriver). No dead flash, and the picker no longer offers an unrunnable driver.
- RmtLedDriver / ParallelLedDriver (LCD + Parlio base): loopbackTxPin / loopbackRxPin moved to int8/addPin, -1 = unset (so GPIO 0 is usable); the unset check is now >= 0.
- LcdLedDriver: clockPin / dcPin moved to int8/addPin (the i80 bus pins); lastClockPin_ / lastDcPin_ follow.
- RotateModifier: corrected the timing comment — the angle step is phaseNum_/64, so a full turn is 16384/speed ms (the prior comment said 256*256/speed, 4× too long).

Tests:
- unit_AudioModule: pin-default assertions updated 0 -> -1 to match the Pin-control unset sentinel.

Docs / CI:
- performance.md: new three-board incremental-cost analysis (classic / S3 / P4) — per-subsystem cost, light/heavy effect bracket across 16²..16K, and the MultiplyModifier compute-down/memory-up tables. Tables are board-columned so later P4/classic re-runs drop straight in. Records the key findings: linear pixel scaling (no fragmentation), the heavy effect is the 16K bottleneck (P4 57 / S3 20 / classic 16 FPS), the modifier roughly halves heavy tick (¼ logical grid) at a LUT-memory cost, and 16K Noise+Multiply runs on the no-PSRAM classic at 35 FPS render-only (confirming it is not a no-PSRAM blocker).
- architecture.md: light-domain rule — an effect renders a pattern, it does not transform geometry; mirror/tile/rotate/scroll/mask belong in a modifier (WLED folds these into effect code; we don't), so any effect composes with any modifier.
- new scenario_perf_light (fast, ≤64², driver-free) and scenario_perf_full (comprehensive: per-subsystem adds, all-drivers-this-board, light/heavy bracket to 16K, modifier sweep). Both mutate + canvas-preparing: clear_children to a clean start, keeping only pre-wired apparatus; disable Audio + Devices so neither pollutes the tick. They subsume and replace scenario_Layer_buildup, scenario_GridLayout_grid_sizes, and scenario_AllEffects_grid_sizes (deleted). scenario-tests.md / unit-tests.md regenerated; testing.md example + backlog references updated.
- backlog: NoiseEffect cost extended with the S3 16K data point; new entries for MultiplyModifier LUT-memory (re-verify the working classic+ArtNet setup, not an impossibility), disabling a module should release its resources not just stop its loop (onEnabledChanged audit), and the future classic I2S driver following the per-chip gating pattern. Removed the now-moot "RipplesEffect in the AllEffects sweep" item.
- .gitignore: ignore the generated test/scenario-baseline.json cache.

Reviews:
- 🐇 RmtLedDriver.md future-work core-1-task sentence -> removed (present tense; it's in leddriver-deferred.md) (fixed)
- 🐇 RandomMapModifier.md "first dynamic modifier" -> "a dynamic modifier" (RotateModifier is also dynamic) (fixed)
- 🐇 RotateModifier.h full-turn timing comment off by 4× -> corrected (fixed)
- 🐇 DevicesModule.md progress range 1..254 -> 0..254 (fixed)
- 🐇 DevicesModule.md two future-work/backlog sentences -> removed (present tense) (fixed)
- 🐇 RmtLedDriver loopback pins used uint16/addUint16 instead of the standard addPin -> migrated to int8/addPin, and extended the same fix to every single-GPIO control (Audio, Lcd, Parlio/Parallel base) for consistency (fixed)

+++ src/light/drivers/ParlioLedDriver.h
@@ -27,8 +27,8 @@ namespace mm {
 class ParlioLedDriver : public ParallelLedDriver<ParlioLedDriver> {
 public:
     // All controls default to UNSET — pins="", ledsPerPin="" (= all lights on the
-    // first lane, even-split with one lane), loopbackRxPin=0 — so no constructor is
-    // needed (the base zero-initialises them). Pins/loopback are unset because the
+    // first lane, even-split with one lane), loopbackRxPin=-1 — so no constructor is
+    // needed (the base default-initialises them). Pins/loopback are unset because the
     // strand is user-soldered: a hard-coded pin would guess the user's wiring and
     // could drive a pin committed elsewhere ("default only when it cannot do harm",
     // see decisions.md). The P4-NANO bench used pins "20,21,22,23,24,25,26,27",
@@ -66,7 +66,8 @@ public:
     // unit on lane 0 (no i80 full-bus workaround needed; no WR/DC to pass).
     platform::RmtLoopbackResult busLoopback(const uint8_t* frame, size_t frameBytes,
                                             size_t dataBytes, uint8_t rowBits) {
-        return platform::parlioWs2812Loopback(laneList_, laneCount_, loopbackRxPin,
+        return platform::parlioWs2812Loopback(laneList_, laneCount_,
+                                              static_cast<uint16_t>(loopbackRxPin),
                                               frame, frameBytes, dataBytes, rowBits);
     }
